SUpport for ASP.NET
FW:updating to find DOTNET version

Fixing logic for version detected & created for dotnot

Support  for Java

+++ src/webapp/azext_webapp/custom.py
@@ -33,7 +33,7 @@ from .create_util import (
     web_client_factory
 )
 
-from ._constants import (NODE_RUNTIME_NAME, OS_DEFAULT)
+from ._constants import (NODE_RUNTIME_NAME, OS_DEFAULT, JAVA_RUNTIME_NAME)
 
 logger = get_logger(__name__)
 
@@ -64,7 +64,8 @@ def create_deploy_webapp(cmd, name, location=None, dryrun=False):
     else:
         sku = lang_details.get("default_sku")
         language = lang_details.get("language")
-        os_val = "Linux" if language.lower() == NODE_RUNTIME_NAME else OS_DEFAULT
+        is_java = language.lower() == JAVA_RUNTIME_NAME
+        os_val = "Linux" if language.lower() == NODE_RUNTIME_NAME or is_java else OS_DEFAULT
         # detect the version
         data = get_runtime_version_details(lang_details.get('file_loc'), language)
         version_used_create = data.get('to_create')
@@ -100,7 +101,6 @@ def create_deploy_webapp(cmd, name, location=None, dryrun=False):
 
     src_path = "{} {}".format(src_dir.replace("\\", "\\\\"), str_no_contents_warn)
     rg_str = "{} {}".format(rg_name, rg_mssg)
-
     dry_run_str = r""" {
             "name" : "%s",
             "serverfarm" : "%s",
@@ -152,32 +152,38 @@ def create_deploy_webapp(cmd, name, location=None, dryrun=False):
         logger.warning("App '%s' already exists", name)
 
     if do_deployment:
-        # setting to build after deployment
-        logger.warning("Updating app settings to enable build after deployment")
-        update_app_settings(cmd, rg_name, name, ["SCM_DO_BUILD_DURING_DEPLOYMENT=true"])
-        # work around until the timeout limits issue for linux is investigated & fixed
-        # wakeup kudu, by making an SCM call
+        if not is_java:
+            # setting to build after deployment
+            logger.warning("Updating app settings to enable build after deployment")
+            update_app_settings(cmd, rg_name, name, ["SCM_DO_BUILD_DURING_DEPLOYMENT=true"])
+            # work around until the timeout limits issue for linux is investigated & fixed
+            # wakeup kudu, by making an SCM call
 
         import requests
         # work around until the timeout limits issue for linux is investigated & fixed
         user_name, password = _get_site_credential(cmd.cli_ctx, rg_name, name)
         scm_url = _get_scm_url(cmd, rg_name, name)
+
         import urllib3
         authorization = urllib3.util.make_headers(basic_auth='{0}:{1}'.format(user_name, password))
         requests.get(scm_url + '/api/settings', headers=authorization)
 
-        logger.warning("Creating zip with contents of dir %s ...", src_dir)
-        # zip contents & deploy
-        zip_file_path = zip_contents_from_dir(src_dir, language)
+        if is_java:
+            zip_file_path = src_path + '\\\\' + lang_details.get('file_loc')[0]
+        else:
+            logger.warning("Creating zip with contents of dir %s ...", src_dir)
+            # zip contents & deploy
+            zip_file_path = zip_contents_from_dir(src_dir, language)
 
         logger.warning("Deploying and building contents to app."
                        "This operation can take some time to finish...")
         enable_zip_deploy(cmd, rg_name, name, zip_file_path)
-        # Remove the file afer deployment, handling exception if user removed the file manually
-        try:
-            os.remove(zip_file_path)
-        except OSError:
-            pass
+        if not is_java:
+            # Remove the file afer deployment, handling exception if user removed the file manually
+            try:
+                os.remove(zip_file_path)
+            except OSError:
+                pass
     else:
         logger.warning("No 'NODE' or 'DOTNETCORE' package detected, skipping zip and deploy process")
     create_json.update({'app_url': url})
@@ -247,3 +253,4 @@ def create_tunnel(cmd, resource_group_name, name, port, slot=None):
                 break
     logger.warning('Tunnel is ready! Creating on port %s', port)
     tunnel_server.start_server()
+
